refactor: Restructure application into smaller components
